Rearranged package and extracted console application to ConversationExporterApp
