+++ .gitignore
@@ -2,3 +2,4 @@
 target/
 bin/
 *.iml
+.json
