+++ pom.xml
@@ -4,13 +4,14 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.mindlinksoft.recruitment.mychat</groupId>
+    <groupId>com.mindlinksoft.recruitment.juliankubelec.mychat</groupId>
     <artifactId>my-chat</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
     <dependencies>
@@ -62,7 +63,7 @@
                     <manifest>
                       <addClasspath>true</addClasspath>
                       <classpathPrefix>lib/</classpathPrefix>
-                      <mainClass>com.mindlinksoft.recruitment.mychat.ConversationExporter</mainClass>
+                      <mainClass>com.mindlinksoft.recruitment.juliankubelec.mychat.ConversationExporter</mainClass>
                     </manifest>
                   </archive>
                 </configuration>
