+++ src/main/java/com/mindlinksoft/recruitment/juliankubelec/mychat/ConversationExporter.java
@@ -1,12 +1,10 @@
-package com.mindlinksoft.recruitment.mychat;
+package com.mindlinksoft.recruitment.juliankubelec.mychat;
 
 import com.google.gson.*;
 
-import com.mindlinksoft.recruitment.mychat.exceptions.EmptyTextFileException;
+import com.mindlinksoft.recruitment.juliankubelec.mychat.exceptions.EmptyTextFileException;
 import picocli.CommandLine;
 import picocli.CommandLine.ParameterException;
-import picocli.CommandLine.ParseResult;
-import picocli.CommandLine.UnmatchedArgumentException;
 
 import java.io.*;
 import java.lang.reflect.Type;
@@ -23,65 +21,7 @@ public class ConversationExporter {
     private String filterKeyword;
     private List<String> blacklist;
     private boolean includeReport = false;
-
-
-
     private CommandLine cmd;
-
-    /**
-     * The application entry point.
-     * @param args The command line arguments.
-     * //@throws Exception Thrown when something bad happens.
-     */
-
-    public static void main(String[] args) {
-        // We use picocli to parse the command line - see https://picocli.info/
-        ConversationExporterConfiguration configuration = new ConversationExporterConfiguration();
-        ConversationExporter exporter = new ConversationExporter();
-        exporter.setCmd(new CommandLine(configuration));
-        CommandLine cmd = exporter.getCmd();
-
-        try {
-            ParseResult parseResult = cmd.parseArgs(args);
-            if (parseResult.isUsageHelpRequested()) {
-                cmd.usage(cmd.getOut());
-                System.exit(cmd.getCommandSpec().exitCodeOnUsageHelp());
-                return;
-            }
-
-            if (parseResult.isVersionHelpRequested()) {
-                cmd.printVersionHelp(cmd.getOut());
-                System.exit(cmd.getCommandSpec().exitCodeOnVersionHelp());
-                return;
-            }
-
-            if(parseResult.hasMatchedOption(configuration.userOpt)) {
-                exporter.setFilterUserId(configuration.filterUserId);
-            }
-            else if(parseResult.hasMatchedOption(configuration.filterKeyword)) {
-                exporter.setFilterKeyword(configuration.filterKeyword);
-            }
-            else if(parseResult.hasMatchedOption(configuration.blacklistOpt)) {
-                exporter.setBlacklist(configuration.blacklist);
-            }else if(parseResult.hasMatchedOption(configuration.reportOpt)){
-                exporter.setIncludeReport(configuration.reportIncluded);
-            }
-
-            exporter.exportConversation(configuration.inputFilePath, configuration.outputFilePath);
-
-            System.exit(cmd.getCommandSpec().exitCodeOnSuccess());
-        } catch (ParameterException ex) {
-            cmd.getErr().println(ex.getMessage());
-            if (!UnmatchedArgumentException.printSuggestions(ex, cmd.getErr())) {
-                ex.getCommandLine().usage(cmd.getErr());
-            }
-
-            System.exit(cmd.getCommandSpec().exitCodeOnInvalidInput());
-        } catch (Exception ex) {
-            ex.printStackTrace(cmd.getErr());
-            System.exit(cmd.getCommandSpec().exitCodeOnExecutionException());
-        }
-    }
 
     /**
      * Exports the conversation at {@code inputFilePath} as JSON to {@code outputFilePath}.
